backend/requirements.txt
# FastAPI and server
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Data processing
pandas==2.1.4
numpy==1.26.3

# HTTP requests and scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
yfinance==0.2.35

# Scheduling
apscheduler==3.10.4

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS and middleware
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Logging
python-json-logger==2.0.7

# Export functionality
reportlab==4.0.8
